Consumer/pyproject.toml
[project]
name = "consumer"
version = "0.1.0"
description = ""
requires-python = ">=3.11"

dependencies = [
    "confluent-kafka",
    "pydantic",
    "psycopg2-binary"
]
